--- src/pages/Register/index.ts
@@ -1,1 +1,0 @@
-export { RegisterPage } from './tmp_Register'
