--- src/pages/Register/steps/index.ts
@@ -1,5 +1,0 @@
-export { Step1 } from './Step1'
-export { Step2 } from './Step2'
-export { Step3 } from './Step3'
-export { Step4 } from './Step4'
-export { Success } from './Success'
